Moving ConfigChangeApplicator to the right module
