+++ dynamic-config/server/configuration-provider/src/main/java/org/terracotta/dynamic_config/server/configuration/service/ConfigChangeApplicator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terracotta.dynamic_config.server.configuration.nomad;
+package org.terracotta.dynamic_config.server.configuration.service;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
